feat(redux-utils): implement logout tests
